change error

+++ demo/src/main/java/project/demo/repository/DocumentRepository.java
@@ -3,5 +3,5 @@ package project.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.demo.entity.Document;
 
-public interface DocumentRepository extends JpaRepository<Document, Long> {
+public interface DocumentRepository extends JpaRepository<Document, Integer> {
 }
